added notfound exceptions

+++ MoodBE/src/com/moodlysis/moodbe/integration/Question.java
@@ -10,6 +10,7 @@ import java.sql.Statement;
 import com.moodlysis.moodbe.DatabaseConnection;
 import com.moodlysis.moodbe.integrationinterfaces.QuestionInterface;
 import com.moodlysis.moodbe.requestexceptions.MoodlysisInternalServerError;
+import com.moodlysis.moodbe.requestexceptions.MoodlysisNotFound;
 
 public class Question implements QuestionInterface {
 
@@ -22,7 +23,7 @@ public class Question implements QuestionInterface {
 	}
 	
 	@Override
-	public questionInfo getQuestion(int questionID) throws MoodlysisInternalServerError {
+	public questionInfo getQuestion(int questionID) throws MoodlysisInternalServerError, MoodlysisNotFound {
 		questionInfo info = new questionInfo();
 		info.questionID = questionID;
 		PreparedStatement questionGet  = null;
@@ -38,7 +39,9 @@ public class Question implements QuestionInterface {
     		questionGet = conn.prepareStatement(query);
     		questionGet.setInt(1, questionID);
 			table = questionGet.executeQuery();
-			table.next();
+			if (!table.next()) {
+				throw new MoodlysisNotFound("Question not found. Question may have expired or been deleted.");
+			}
 			formID = table.getInt("FormID");
 			numInForm = table.getInt("NumInForm");
 			type = table.getString("Type");
@@ -147,11 +150,19 @@ public class Question implements QuestionInterface {
 	}
 
 	@Override
-	public boolean editQuestionDetails(int questionID, String questionType, String text, String options) throws MoodlysisInternalServerError {
+	public boolean editQuestionDetails(int questionID, String questionType, String text, String options) throws MoodlysisInternalServerError, MoodlysisNotFound {
 		// TODO fix so if error occurs return false but still execute finally statement
-		PreparedStatement questionEdit  = null;
+		PreparedStatement questionEdit  = null;PreparedStatement existCheck = null;
+		ResultSet table = null;
 		try {
 			conn.setAutoCommit(false);
+			String selectQuery = "SELECT * FROM QUESTIONS WHERE QuestionID = ?";
+			existCheck = conn.prepareStatement(selectQuery);
+			existCheck.setInt(1, questionID);
+			table = existCheck.executeQuery();
+			if (!table.next()) {
+				throw new MoodlysisNotFound("Question not found. Question may have expired or been deleted.");
+			}
     		String query = "UPDATE QUESTIONS SET Type = ?, Content = ?, Options = ? WHERE QuestionID = ?";
     		questionEdit = conn.prepareStatement(query);
     		questionEdit.setString(1, questionType);
@@ -175,6 +186,12 @@ public class Question implements QuestionInterface {
 				if (questionEdit != null) {
 					questionEdit.close();
 				}
+				if (existCheck != null) {
+					existCheck.close();
+				}
+				if (table != null) {
+					table.close();
+				}
 			} catch(SQLException e) {
 				e.printStackTrace(this.writer);
 				throw new MoodlysisInternalServerError(e.toString());
@@ -183,13 +200,23 @@ public class Question implements QuestionInterface {
 		return true;
 	}
 	
-	public boolean editQuestionPosition(int questionID, int previousID) throws MoodlysisInternalServerError {
+	@Override
+	public boolean editQuestionPosition(int questionID, int previousID) throws MoodlysisInternalServerError, MoodlysisNotFound {
 		PreparedStatement questionEdit  = null;
 		PreparedStatement getNumInForm = null;
 		ResultSet table = null;
+		PreparedStatement existCheck = null;
+		ResultSet table2 = null;
 		int numInForm = -1;
 		try {
 			conn.setAutoCommit(false);
+			String selectQuery = "SELECT * FROM QUESTIONS WHERE QuestionID = ?";
+			existCheck = conn.prepareStatement(selectQuery);
+			existCheck.setInt(1, questionID);
+			table2 = existCheck.executeQuery();
+			if (!table2.next()) {
+				throw new MoodlysisNotFound("Question not found. Question may have expired or been deleted.");
+			}
 			if (previousID > 0) {
 				//Query makes sure forms of previous id and given question id are the same
 				//e.g. SELECT QUESTIONS.QuestionID, QUESTIONS.NumInForm FROM (SELECT * FROM QUESTIONS WHERE QuestionID = 4) AS QUESTION INNER JOIN QUESTIONS ON (QUESTION.FormID = QUESTIONS.FormID) WHERE QUESTIONS.QuestionID = 5 OR QUESTIONS.QuestionID = 4 ORDER BY QUESTIONS.NumInForm DESC;
@@ -248,6 +275,12 @@ public class Question implements QuestionInterface {
 				if (table != null) {
 					table.close();
 				}
+				if (existCheck != null) {
+					existCheck.close();
+				}
+				if (table2 != null) {
+					table2.close();
+				}
 			} catch(SQLException e) {
 				e.printStackTrace(this.writer);
 				throw new MoodlysisInternalServerError(e.toString());
@@ -257,10 +290,19 @@ public class Question implements QuestionInterface {
 	}
 
 	@Override
-	public boolean deleteQuestion(int questionID) throws MoodlysisInternalServerError {
+	public boolean deleteQuestion(int questionID) throws MoodlysisInternalServerError, MoodlysisNotFound {
 		PreparedStatement questionDelete  = null;
+		PreparedStatement existCheck = null;
+		ResultSet table = null;
 		try {
 			conn.setAutoCommit(false);
+			String selectQuery = "SELECT * FROM QUESTIONS WHERE QuestionID = ?";
+			existCheck = conn.prepareStatement(selectQuery);
+			existCheck.setInt(1, questionID);
+			table = existCheck.executeQuery();
+			if (!table.next()) {
+				throw new MoodlysisNotFound("Question not found. Question may have expired or been deleted.");
+			}
     		String query = "DELETE FROM QUESTIONS WHERE QuestionID = ?";
     		questionDelete = conn.prepareStatement(query);
     		questionDelete.setInt(1, questionID);
@@ -280,6 +322,12 @@ public class Question implements QuestionInterface {
 			try {
 				if (questionDelete != null) {
 					questionDelete.close();
+				}
+				if (existCheck != null) {
+					existCheck.close();
+				}
+				if (table != null) {
+					table.close();
 				}
 			} catch(SQLException e) {
 				e.printStackTrace(this.writer);
